recipe-app-backend/requirements.txt
pytest~=7.4.4
h11~=0.14.0
pip~=24.0
toml~=0.10.2
wheel~=0.43.0
brotli~=1.0.9
cryptography~=42.0.5
lxml~=5.3.0
tornado~=6.4.1
keyring~=24.3.1
pytz~=2024.1
requests~=2.32.3
dask~=2024.5.0
rich~=13.8.0
colorama~=0.4.6
ipywidgets~=7.8.1
ipython~=8.25.0
notebook~=7.0.8
idna~=3.8
multidict~=6.0.4
sniffio~=1.3.0
attrs~=24.2.0
click~=8.1.7
httpcore~=1.0.2
pygments~=2.18.0
zstandard~=0.22.0
certifi~=2024.8.30
jiter~=0.6.1
numpy~=1.26.4
setuptools~=69.5.1
overrides~=7.4.0
cffi~=1.16.0
numba~=0.60.0
mypy~=1.10.0
distro~=1.9.0
openai~=1.52.1
aiohttp~=3.9.5
pandas~=2.2.2
matplotlib~=3.8.4
plotly~=5.22.0
scipy~=1.14.1
scikit-learn~=1.4.2
tenacity~=8.5.0
openpyxl~=3.1.2
fsspec~=2024.3.1
pyarrow~=17.0.0
sqlalchemy~=2.0.34
markupsafe~=2.1.5
numexpr~=2.8.7
python-dateutil~=2.9.0post0
wcwidth~=0.2.5
pluggy~=1.0.0
iniconfig~=1.1.1
pexpect~=4.8.0
twisted~=23.10.0
yarl~=1.9.3
aiosignal~=1.2.0
frozenlist~=1.4.0
fastapi~=0.112.2
starlette~=0.38.2
pydantic~=2.5.3
anyio~=4.2.0
ujson~=5.10.0
urllib3~=2.2.2
uvicorn~=0.20.0
six~=1.16.0
greenlet~=3.0.1
psutil~=5.9.0
chardet~=4.0.0
packaging~=24.1
itsdangerous~=2.2.0
jinja2~=3.1.4
httpx~=0.27.0
docutils~=0.18.1
pyodbc~=5.0.1
python-dotenv~=0.21.0